Use BigFloat parsing for Float64 to fix Windows bug. (#143)
* Use BigFloat parsing for Float64 to fix Windows bug.

Remove unnecessary `@eval`.

* Fix atomic for Windows

* Specific atomic for Irrational

* Specific atomic for Rational

* Do correct float rounding in new parse

* Bump version of CRlibm required

* Int -> Int64

+++ test/interval_tests/construction.jl
@@ -106,7 +106,7 @@ using Base.Test
     @test b == @floatinterval(pi)
     @test nextfloat(b.lo) == b.hi
     @test convert(Interval{Float64}, @biginterval(0.1)) == a
-    x = typemax(Int)
+    x = typemax(Int64)
     @test @interval(x) == @floatinterval(x)
     @test isthin(@interval(x)) == false
     x = rand()
